changed stuff from long long to intmax_t

+++ src/ft_lltoa_flag.c
@@ -6,13 +6,13 @@
 /*   By: cterblan <cterblan@student.wethinkcode>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/06/06 11:14:38 by cterblan          #+#    #+#             */
-/*   Updated: 2018/08/23 07:38:28 by cterblan         ###   ########.fr       */
+/*   Updated: 2018/08/23 10:48:11 by cterblan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../inc/ft_printf.h"
 
-static size_t	ft_getlen(long long n)
+static size_t	ft_getlen(intmax_t n)
 {
 	size_t	i;
 
@@ -22,16 +22,16 @@ static size_t	ft_getlen(long long n)
 	return (i);
 }
 
-static void		ft_iszero(char *new, long long n)
+static void		ft_iszero(char *new, intmax_t n)
 {
 	if (n == 0)
 		new[0] = '0';
 }
 
-char			*ft_lltoa_flag(long long n)
+char			*ft_lltoa_flag(intmax_t n)
 {
 	size_t				i;
-	unsigned long long	tmp;
+	uintmax_t	tmp;
 	char				*new;
 
 	i = ft_getlen(n);
