+++ src/ft_ulltoa.c
@@ -6,7 +6,7 @@
 /*   By: cterblan <cterblan@student.wethinkcode>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/08/22 07:33:44 by cterblan          #+#    #+#             */
-/*   Updated: 2018/08/22 08:29:19 by cterblan         ###   ########.fr       */
+/*   Updated: 2018/08/23 10:49:13 by cterblan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 #include "../inc/ft_printf.h"
 
-static size_t	ft_getlen(long long unsigned n, int base)
+static size_t	ft_getlen(uintmax_t n, int base)
 {
 	size_t	i;
 
@@ -24,16 +24,16 @@ static size_t	ft_getlen(long long unsigned n, int base)
 	return (i);
 }
 
-static void		ft_iszero(char *new, long long unsigned n)
+static void		ft_iszero(char *new, uintmax_t n)
 {
 	if (n == 0)
 		new[0] = '0';
 }
 
-char			*ft_ulltoa_base(long long unsigned n, int base)
+char			*ft_ulltoa_base(uintmax_t n, int base)
 {
 	size_t				i;
-	unsigned long long	tmp;
+	uintmax_t	tmp;
 	char				*new;
 
 	i = ft_getlen(n, base);
